port SHA-1: 7b1e1ccadb83ce8f207838d500edf12f8effc856
* Fix some warnings from new inspection tool


git-tfs-id: [https://tfs.codeplex.com/tfs/TFS05/]$/ZXingNet/trunk;C90069

+++ Source/lib/pdf417/encoder/PDF417HighLevelEncoder.cs
@@ -139,22 +139,22 @@ namespace ZXing.PDF417.Internal
          //Construct inverse lookups
          for (int idx = 0; idx < MIXED.Length; idx++)
             MIXED[idx] = -1;
-         for (sbyte i = 0; i < TEXT_MIXED_RAW.Length; i++)
+         for (int i = 0; i < TEXT_MIXED_RAW.Length; i++)
          {
             sbyte b = TEXT_MIXED_RAW[i];
             if (b > 0)
             {
-               MIXED[b] = i;
+               MIXED[b] = (sbyte)i;
             }
          }
          for (int idx = 0; idx < PUNCTUATION.Length; idx++)
             PUNCTUATION[idx] = -1;
-         for (sbyte i = 0; i < TEXT_PUNCTUATION_RAW.Length; i++)
+         for (int i = 0; i < TEXT_PUNCTUATION_RAW.Length; i++)
          {
             sbyte b = TEXT_PUNCTUATION_RAW[i];
             if (b > 0)
             {
-               PUNCTUATION[b] = i;
+               PUNCTUATION[b] = (sbyte)i;
             }
          }
       }
